lectures/08._dotenv/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["William Omø <williamjean.k.o@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
